# services/api/openapi/spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Águia Gestão Veicular API",
    "version": "0.1.0",
    "description": "API REST da plataforma Águia — gestão veicular, SaaS multi-tenant e painel master.",
    "contact": {
      "name": "ÁguiaAuto"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Desenvolvimento" },
    { "url": "https://app.seudominio.com/api", "description": "Produção (via Nginx /api)" }
  ],
...
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login cliente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Tokens JWT" }, "401": { "description": "Credenciais inválidas" } }
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Perfil do cliente autenticado",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Usuário" }, "401": { "description": "Não autenticado" } }
      }
    },
    "/v1/admin/auth/login": {
      "post": {
        "tags": ["Admin Auth"],
        "summary": "Login admin (cookies HttpOnly + CSRF)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Sessão admin" }, "401": { "description": "Credenciais inválidas" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token cliente (Authorization: Bearer)"
      },
      "adminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_access",
        "description": "Sessão admin HttpOnly + header X-CSRF-Token em mutações"
      }
    },
...
  }
}